Initial commit.
It begins. Err… not really. I have been working on this project for
several months in a private GitHub repository and am working on making
it public now.

I had two options for how I would accomplish this:

1. Simply change my existing GitHub repository from being “private” to
being “public”.
2. Create a new repository and move the work over, piece by piece.

Obviously, I’ve decided to choose the more difficult option, and for
two reasons:

1. I, like most developers, have a fear of inadequacy about my code
quality and want to prevent the flaws in my code from being exposed.
That’s the bad reason.
2. I suspect that I have various pieces of personal information, such
as private keys, stored in the private repository which could cause
damage if publicly released. It would be an arduous task to mine
through each commit to permanently erase them. Also, that somewhat
defeats the point of having a permanent git history.

So, I am starting with a “Hello world” message to begin the slow
process of moving my code into the bright light of day so that it can
be examined and critiqued by the public eye.

It will help me deal with that insecurity issue I mentioned.