Created the Model.
The `Model` will be the base class for all objects that have data
stored in the Google App Engine datastore. The `Model` will have
methods for basic CRUD access to the datastore.

The creation of the JUnit test case for `Model` adds JUnit to the build
process, but there are currently no plans to hook the test process into
the build process.