The Write page now displays differently based on URL.
The Write page has three *modes*:

1. If the `from` parameter is “0”, the write page is in “new story”
mode, which displays a title text input.
2. If the `from` parameter is a number greater than 0 and the
`linkIndex` is between 0 and 5 (exclusive), the write page is in “valid
option” mode, which displays the option text and write form. The `from`
parameter should actually be a `PageId` with a valid, non-null version
component.
3. Else, the write page is in “invalid option” mode, in which case an
error message is displayed to the user.