Built the controller for saving new stories.
I split the `write.jsp` page out into a `new.jsp` page which only
handles the creation of new stories. The JSP page passes form data to
the `SubmitNewStoryServlet` which creates a `SubmitNewStoryController`
to build the `StoryBeginning`, `StoryPage` and `StoryOption` objects.
These objects are then all put in the datastore, and the user is
redirected to the `read.jsp` page of the page they just wrote.