Added the option text to the write page.
When a user has clicked on a link to extend a story, it is helpful to
have the text of the option that they selected visible on the page.
This will make it easier for the writer to remember what they were
intending to write.

This feature was recommended by Nic Henwood.