Added a user preferences page.
The user can now change their pen name.

There is currently a bug where the pen name change fails if there are
multiple copies of the user in the datastore. There shouldn’t be
multiple copies of the user in the datastore, but currently the
creation process does it sometimes.