Fixed NullPointerException when no followers.
Resolves issue #208.

When the user is logged in but has no followers, the follower list is `null`.

This was causing a `NullPointerException` when notifications were sent to all
followers. A for-each loop tried to iterate over the `null` list. Now, if the
list is `null`, the loop is skipped, because the list is empty.