Dealt with Russian spam bots.
A few days ago, bots started spamming Dendrite by rewriting hundreds of
pages and starting hundreds of new stories. Dendrite usually stays
within the free quota of Google App Engine, so when the bots started
making loads of requests to the datastore, I started getting charged.
I turned off the billing and Dendrite went down for a day. I added
`dos.xml` to block the subnets which the bot requests originated from.
Once I confirmed that the bots were getting 403 FORBIDDEN responses to
their HTTP requests, I built a tool called `BotCleaner` to erase large
numbers of alternative versions for pages that the bots had written. It
worked in my test environment but not on the live, public system. This
meant that I had to spend an hour or so manually deleting large sets of
pages and stories using the Google Cloud console in my browser.

A tool should be written to flag suspicious submissions and prep them in
batches for mass deletion once verified by a human as being all spam.